Move all errors into a common submodule
