+++ gryf/src/adapt/undirect.rs
@@ -12,8 +12,10 @@ use gryf_derive::{
 
 // TODO: Remove these imports once hygiene of procedural macros is fixed.
 use crate::core::{
-    marker::EdgeType, AddEdgeError, AddVertexError, Edges, EdgesBase, EdgesBaseWeak, EdgesMut,
-    EdgesWeak, GraphBase, Guarantee, Vertices, VerticesBaseWeak, VerticesMut, VerticesWeak,
+    error::{AddEdgeError, AddVertexError},
+    marker::EdgeType,
+    Edges, EdgesBase, EdgesBaseWeak, EdgesMut, EdgesWeak, GraphBase, Guarantee, Vertices,
+    VerticesBaseWeak, VerticesMut, VerticesWeak,
 };
 
 #[derive(
