+++ gryf/src/core.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod facts;
 pub mod id;
 pub mod marker;
